pyproject.toml
[build-system]
requires = [
    "setuptools", 
    "setuptools_scm",
]
build-backend = "setuptools.build_meta"

[project]
name = "robotodo"
dynamic = ["version"]
requires-python = ">=3.11"
readme = "README.md"
license = {file = "LICENSE.txt"}
dependencies = [
    # TODO
    "bracex",
    "wcmatch",
    # TODO
    "einops",
    "scipy",

    # TODO tensorspecs
    "jax",
    "jaxtyping",
]

[project.optional-dependencies]
engines_isaac = [
    # TODO 
    "torch",
    # TODO
    "nest_asyncio",
    "isaacsim[all,extscache]>=5.0.0,<=6.0.0",
    "warp-lang",
]
algos_pi0 = [
    # TODO mv pi0
    "jax",
    "jaxtyping",
    "optax",
    "einops",
    "flax!=0.11,<0.13.0",
    "fsspec[gcs]>=2024.6.0",
    "orbax-checkpoint>=0.11.13",
    "sentencepiece",
    "platformdirs",
]
algos_pi0_test = [
    "openpi-client @ git+https://github.com/Physical-Intelligence/openpi.git#subdirectory=packages/openpi-client",
    "openpi @ git+https://github.com/Physical-Intelligence/openpi.git",
    "pytest",
]
algos_motion_planning = [    
    # TODO
    # TODO cc1plus error: conda install -c conda-forge cxx-compiler
    "ninja", # NOTE required for jit compilation of curobo torch exts
    "nvidia_curobo @ git+https://github.com/NVlabs/curobo.git@v0.7.7",
]
test = [
    "pytest", 
]
lint = [
    "ruff",
]
docs = [
    "jupyter-book",
    "sphinxcontrib-mermaid",
]

[tool.setuptools.packages.find]
where = ["packages"]
namespaces = true

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = ["packages"]
addopts = "--import-mode=importlib --doctest-modules"

[tool.ruff]
line-length = 80
indent-width = 4
target-version = "py311"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
